pyproject.toml
[project]
name = "jira-reviewer"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = ["mcp[cli]>=1.0.0", "python-dotenv>=1.0.0", "pyyaml>=6.0.0"]

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[project.scripts]
jira-reviewer-mcp = "jira_reviewer.server:main"
